+++ Cargo.toml
@@ -94,7 +94,6 @@ webrogue-hub-client-openapi = { path = "crates/hub-client/openapi", default-feat
 webrogue-icons = { path = "crates/icons" }
 webrogue-gfx = { path = "crates/gfx" }
 webrogue-gfx-winit = { path = "crates/gfx-winit" }
-# webrogue-gfx-fallback = { path = "crates/gfx-fallback" }
 webrogue-gfxstream = { path = "crates/gfxstream" }
 webrogue-gfxstream-lib  = { path = "crates/gfxstream-lib" }
 webrogue-lld = { path = "crates/lld" }
@@ -103,7 +102,7 @@ webrogue-wrapp = { path = "crates/wrapp" }
 webrogue-launcher = { path = "crates/launcher" }
 webrogue-launcher-server-openapi = { path = "crates/launcher/server-openapi", default-features = false }
 
-# Wasmtime
+# Wasmtime crates
 wasmtime = { path = "external/wasmtime/crates/wasmtime", default-features = false }
 wiggle = { path = "external/wasmtime/crates/wiggle", default-features = false }
 wasi-common = { path = "external/wasmtime/crates/wasi-common", default-features = false }
@@ -112,6 +111,7 @@ wiggle-generate = { path = "external/wasmtime/crates/wiggle/generate", default-f
 wasmtime-wasi-threads = { path = "external/wasmtime/crates/wasi-threads" }
 wasmtime-internal-debugger = { path = "external/wasmtime/crates/debugger" }
 
+# other crates
 image = { version = "0.25", default-features = false }
 imageproc = { version = "0.26", default-features = false }
 object = { version = "0.38", default-features = false }
@@ -174,14 +174,12 @@ ico = "0.5.0"
 tar = "0.4"
 gdbstub = "0.7.10"
 gdbstub_arch = "0.3.3"
-wasm_thread = { version = "0.3.3", default-features = false }
 wasm-bindgen = "0.2.114"
 wasm-bindgen-futures = "0.4"
 web-sys = "0.3.91"
 parking_lot = "0.12.5"
 getrandom = { version = "0.4.2", default-features = false }
 cap-rand = { version = "3.4.5", features = ["small_rng"] }
-
 reqwest = { version = "0.13" }
 zip-extract = { version = "0.4.1" }
 
@@ -225,5 +223,3 @@ codegen-units = 1
 [patch.crates-io]
 zstd-safe = { path = "external/zstd-rs/zstd-safe" }
 zstd = { path = "external/zstd-rs" }
-wasm_thread = { path = "external/wasm_thread" }
-# softbuffer = { git = "https://github.com/rust-windowing/softbuffer.git", rev = "7140e44" }
